+++ app/src/main/java/com/asfoundation/wallet/ui/iab/ExpressCheckoutBuyView.java
@@ -1,0 +1,8 @@
+package com.asfoundation.wallet.ui.iab;
+
+/**
+ * Created by franciscocalado on 20/07/2018.
+ */
+
+public interface ExpressCheckoutBuyView {
+}
