+++ app/src/main/res/layout/fragment_iab.xml
@@ -9,11 +9,11 @@
     android:background="@drawable/background_card"
     >
 
-  <include layout="@layout/iab_activity_loading"/>
-  <include layout="@layout/iab_activity_transaction_completed"/>
-  <include layout="@layout/iab_activity_error"/>
+  <include layout="@layout/fragment_iab_loading"/>
+  <include layout="@layout/fragment_iab_transaction_completed"/>
+  <include layout="@layout/fragment_iab_error"/>
   <include
-      layout="@layout/iab_activity_buy"
+      layout="@layout/fragment_iab_buy"
       />
 
   <include
