+++ app/src/main/res/layout/fragment_iab_buy.xml
@@ -7,10 +7,10 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@drawable/background_card"
+    android:elevation="2dp"
     android:minHeight="160dp"
     android:padding="16dp"
-    android:elevation="2dp"
-    tools:showIn="@layout/iab_activity"
+    tools:showIn="@layout/fragment_iab"
     >
 
   <ImageView
