+++ app/src/main/res/layout/fragment_iab_error.xml
@@ -14,7 +14,7 @@
     app:layout_constraintEnd_toEndOf="parent"
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintTop_toTopOf="parent"
-    tools:showIn="@layout/iab_activity"
+    tools:showIn="@layout/fragment_iab"
     >
 
   <TextView
