+++ app/src/main/res/layout/fragment_iab_transaction_completed.xml
@@ -14,7 +14,7 @@
     app:layout_constraintEnd_toEndOf="parent"
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintTop_toTopOf="parent"
-    tools:showIn="@layout/iab_activity"
+    tools:showIn="@layout/fragment_iab"
     >
   <TextView
       android:id="@+id/activity_iab_transaction_completed_title"
